// next/app/dashboard/subscribers/add/z-schema.tsx
import { z } from 'zod'

export const schema = z.object({
    id: z.string().optional(), // Optional because it will be auto-generated by MongoDB
    email: z.string().email(), // Ensures the email is in the correct format
    name: z.string().min(1, {
        message: 'Name is required'
    }), // Optional field
})